reddit_producer/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
kafka-python==2.0.1
praw==7.1.0
prawcore==1.4.0
requests==2.24.0
six==1.15.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
